+++ postinstall.sh
@@ -1,3 +1,3 @@
-. ci/setup.sh
+. ./init.sh
 ts-node tools/update-schematics.ts
 cd tools && yarn