+++ src/app/docs/components/button-demo/icon-label-buttons/icon-label-buttons.component.ts
@@ -1,28 +1,34 @@
 import { Component } from '@angular/core';
-import { LyTheme2 } from '@alyle/ui';
+import { StyleRenderer, ThemeVariables, lyl } from '@alyle/ui';
 
-const STYLES = () => ({
-  button: {
-    margin: '1em'
-  },
-  labelBefore: {
-    paddingAfter: '8px'
-  },
-  labelAfter: {
-    paddingBefore: '8px'
-  },
-  iconSmall: {
-    fontSize: '20px'
-  }
-});
+const STYLES = (theme: ThemeVariables) => {
+  const { before, after } = theme;
+  return {
+    button: lyl `{
+      margin: 1em
+    }`,
+    labelBefore: lyl `{
+      padding-${after}: 8px
+    }`,
+    labelAfter: lyl `{
+      padding-${before}: 8px
+    }`,
+    iconSmall: lyl `{
+      font-size: 20px
+    }`
+  };
+};
 
 @Component({
   selector: 'aui-icon-label-buttons',
-  templateUrl: './icon-label-buttons.component.html'
+  templateUrl: './icon-label-buttons.component.html',
+  providers: [
+    StyleRenderer
+  ]
 })
 export class IconLabelButtonsComponent {
-  readonly classes = this._theme.addStyleSheet(STYLES);
+  readonly classes = this.sRenderer.renderSheet(STYLES);
   constructor(
-    private _theme: LyTheme2
+    readonly sRenderer: StyleRenderer
   ) { }
 }