+++ src/app/docs/components/snack-bar-demo/basic-snack-bar/basic-snack-bar.component.ts
@@ -6,6 +6,9 @@ import { LySnackBarDismiss } from '@alyle/ui/snack-bar';
   templateUrl: './basic-snack-bar.component.html'
 })
 export class BasicSnackBarComponent {
+  data = {
+    val: 'test'
+  };
   afterDismissed(e: LySnackBarDismiss) {
     console.log(e);
   }