+++ src/lib/image-cropper/image-cropper.ts
@@ -10,11 +10,15 @@ import {
   Renderer2,
   HostListener,
   OnDestroy,
-  NgZone
+  NgZone,
+  Inject
 } from '@angular/core';
-import { LyTheme2, mergeDeep, LY_COMMON_STYLES, ThemeVariables, lyl, ThemeRef, StyleCollection, LyClasses, StyleTemplate } from '@alyle/ui';
+import { mergeDeep, LY_COMMON_STYLES, ThemeVariables, lyl, ThemeRef, StyleCollection, LyClasses, StyleTemplate, StyleRenderer } from '@alyle/ui';
 import { Subject, Observable } from 'rxjs';
 import { take, takeUntil } from 'rxjs/operators';
+import { normalizePassiveListenerOptions } from '@angular/cdk/platform';
+import { DOCUMENT } from '@angular/common';
+import { resizeCanvas } from './resize-canvas';
 
 export interface LyImageCropperTheme {
   /** Styles for Image Cropper Component */
@@ -26,6 +30,7 @@ export interface LyImageCropperVariables {
   cropper?: LyImageCropperTheme;
 }
 
+const activeEventOptions = normalizePassiveListenerOptions({passive: false});
 const STYLE_PRIORITY = -2;
 
 export const STYLES = (theme: ThemeVariables & LyImageCropperVariables, ref: ThemeRef) => {
@@ -60,6 +65,7 @@ export const STYLES = (theme: ThemeVariables & LyImageCropperVariables, ref: The
       top: 0
       left: 0
       display: flex
+      touch-action: none
       & > canvas {
         display: block
       }
@@ -117,7 +123,11 @@ export class ImgCropperConfig {
   type?: string;
   /** Background color( default: null), if is null in png is transparent but not in jpg. */
   fill?: string | null;
-  /** Set anti-aliased( default: true) */
+  /**
+   * Set anti-aliased( default: true)
+   * @deprecated this is not necessary as the cropper will automatically resize the image
+   * to the best quality
+   */
   antiAliased?: boolean = true;
   autoCrop?: boolean;
   output?: ImgOutput | ImgResolution = ImgResolution.Default;
@@ -211,7 +221,10 @@ interface ImgRect {
   changeDetection: ChangeDetectionStrategy.OnPush,
   preserveWhitespaces: false,
   selector: 'ly-img-cropper, ly-image-cropper',
-  templateUrl: 'image-cropper.html'
+  templateUrl: 'image-cropper.html',
+  providers: [
+    StyleRenderer
+  ]
  })
 export class LyImageCropper implements OnDestroy {
   static readonly и = 'LyImageCropper';
@@ -219,7 +232,7 @@ export class LyImageCropper implements OnDestroy {
    * styles
    * @docs-private
    */
-  readonly classes = this.theme.addStyleSheet(STYLES);
+  readonly classes = this.sRenderer.renderSheet(STYLES, true);
   _originalImgBase64?: string;
   private _fileName: string | null;
 
@@ -238,6 +251,13 @@ export class LyImageCropper implements OnDestroy {
   private _imgRect: ImgRect = {} as any;
   private _rotation: number;
   private _sizeInBytes: number | null;
+  private _isSliding: boolean;
+  /** Keeps track of the last pointer event that was captured by the crop area. */
+  private _lastPointerEvent: MouseEvent | TouchEvent | null;
+  private _startPointerEvent: {
+    x: number
+    y: number
+  } | null;
 
   /**
    * When is loaded image
@@ -300,19 +320,36 @@ export class LyImageCropper implements OnDestroy {
   /** Emits whenever the component is destroyed. */
   private readonly _destroy = new Subject<void>();
 
+  /** Used to subscribe to global move and end events */
+  protected _document: Document;
+
   constructor(
+    readonly sRenderer: StyleRenderer,
     private _renderer: Renderer2,
-    private theme: LyTheme2,
     private elementRef: ElementRef<HTMLElement>,
     private cd: ChangeDetectorRef,
-    private _ngZone: NgZone
+    private _ngZone: NgZone,
+    @Inject(DOCUMENT) _document: any,
   ) {
-    this._renderer.addClass(elementRef.nativeElement, this.classes.root);
+    this._document = _document;
+  }
+
+  ngOnInit() {
+    this._ngZone.runOutsideAngular(() => {
+      const element = this._imgContainer.nativeElement;
+      element.addEventListener('mousedown', this._pointerDown, activeEventOptions);
+      element.addEventListener('touchstart', this._pointerDown, activeEventOptions);
+    });
   }
 
   ngOnDestroy() {
     this._destroy.next();
     this._destroy.complete();
+    const element = this._imgContainer.nativeElement;
+    this._lastPointerEvent = null;
+    this._removeGlobalEvents();
+    element.removeEventListener('mousedown', this._pointerDown, activeEventOptions);
+    element.removeEventListener('touchstart', this._pointerDown, activeEventOptions);
   }
 
   private _imgLoaded(imgElement: HTMLImageElement) {
@@ -452,11 +489,7 @@ export class LyImageCropper implements OnDestroy {
           top: originPosition.yc
         };
         this._setStylesForContImg({});
-        this._move({
-          srcEvent: {},
-          deltaX: 0,
-          deltaY: 0
-        });
+        this._simulatePointerMove();
       } else {
         return;
       }
@@ -508,66 +541,111 @@ export class LyImageCropper implements OnDestroy {
     this.setScale(this.minScale);
   }
 
-  _moveStart() {
-    this.offset = {
-      x: this._imgRect.x,
-      y: this._imgRect.y,
-      left: this._imgRect.xc,
-      top: this._imgRect.yc
-    };
-  }
-  _move(event: { srcEvent?: {}; deltaX: any; deltaY: any; }) {
-    let x: number | undefined, y: number | undefined;
-    const canvas = this._imgCanvas.nativeElement;
-    const scaleFix = this._scal3Fix;
-    const config = this.config;
-    const startP = this.offset;
-    if (!scaleFix || !startP) {
+  private _pointerDown = (event: TouchEvent | MouseEvent) => {
+    // Don't do anything if the
+    // user is using anything other than the main mouse button.
+    if (this._isSliding || (!isTouchEvent(event) && event.button !== 0)) {
       return;
     }
 
-    const isMinScaleY = canvas.height * scaleFix < config.height && config.extraZoomOut;
-    const isMinScaleX = canvas.width * scaleFix < config.width && config.extraZoomOut;
-
-    const limitLeft = (config.width / 2 / scaleFix) >= startP.left - (event.deltaX / scaleFix);
-    const limitRight = (config.width / 2 / scaleFix) + (canvas.width) - (startP.left - (event.deltaX / scaleFix)) <= config.width / scaleFix;
-    const limitTop = ((config.height / 2 / scaleFix) >= (startP.top - (event.deltaY / scaleFix)));
-    const limitBottom = (((config.height / 2 / scaleFix) + (canvas.height) - (startP.top - (event.deltaY / scaleFix))) <= (config.height / scaleFix));
-
-    // Limit for left
-    if ((limitLeft && !isMinScaleX) || (!limitLeft && isMinScaleX)) {
-      x = startP.x + (startP.left) - (config.width / 2 / scaleFix);
-    }
-
-    // Limit for right
-    if ((limitRight && !isMinScaleX) || (!limitRight && isMinScaleX)) {
-      x = startP.x + (startP.left) + (config.width / 2 / scaleFix) - canvas.width;
-    }
-
-    // Limit for top
-    if ((limitTop && !isMinScaleY) || (!limitTop && isMinScaleY)) {
-      y = startP.y + (startP.top) - (config.height / 2 / scaleFix);
-    }
-
-    // Limit for bottom
-    if ((limitBottom && !isMinScaleY) || (!limitBottom && isMinScaleY)) {
-      y = startP.y + (startP.top) + (config.height / 2 / scaleFix) - canvas.height;
-    }
-
-    // When press shiftKey, deprecated
-    // if (event.srcEvent && event.srcEvent.shiftKey) {
-    //   if (Math.abs(event.deltaX) === Math.max(Math.abs(event.deltaX), Math.abs(event.deltaY))) {
-    //     y = this.offset.top;
-    //   } else {
-    //     x = this.offset.left;
-    //   }
-    // }
-
-    if (x === void 0) { x = (event.deltaX / scaleFix) + (startP.x); }
-    if (y === void 0) { y = (event.deltaY / scaleFix) + (startP.y); }
-    this._setStylesForContImg({
-      x, y
+    this._ngZone.run(() => {
+      this.offset = {
+        x: this._imgRect.x,
+        y: this._imgRect.y,
+        left: this._imgRect.xc,
+        top: this._imgRect.yc
+      };
+      this._isSliding = true;
+      this._lastPointerEvent = event;
+      this._startPointerEvent = getGesturePointFromEvent(event);
+      event.preventDefault();
+      this._bindGlobalEvents(event);
     });
+
+  }
+
+  /**
+   * Simulate pointerMove with clientX = 0 and clientY = 0,
+   * this is used by `setScale` and `rotate`
+   */
+  private _simulatePointerMove() {
+    this._isSliding = true;
+    this._startPointerEvent = {
+      x: 0,
+      y: 0
+    };
+    this._pointerMove({
+      clientX: 0,
+      clientY: 0,
+      preventDefault: () => {}
+    } as MouseEvent);
+    this._isSliding = false;
+    this._startPointerEvent = null;
+  }
+
+  /**
+   * Called when the user has moved their pointer after
+   * starting to drag.
+   */
+  private _pointerMove = (event: TouchEvent | MouseEvent) => {
+    if (this._isSliding) {
+      event.preventDefault();
+      this._lastPointerEvent = event;
+      let x: number | undefined, y: number | undefined;
+      const canvas = this._imgCanvas.nativeElement;
+      const scaleFix = this._scal3Fix;
+      const config = this.config;
+      const startP = this.offset;
+      const point = getGesturePointFromEvent(event);
+      const deltaX = point.x - this._startPointerEvent!.x;
+      const deltaY = point.y - this._startPointerEvent!.y;
+      if (!scaleFix || !startP) {
+        return;
+      }
+
+      const isMinScaleY = canvas.height * scaleFix < config.height && config.extraZoomOut;
+      const isMinScaleX = canvas.width * scaleFix < config.width && config.extraZoomOut;
+
+      const limitLeft = (config.width / 2 / scaleFix) >= startP.left - (deltaX / scaleFix);
+      const limitRight = (config.width / 2 / scaleFix) + (canvas.width) - (startP.left - (deltaX / scaleFix)) <= config.width / scaleFix;
+      const limitTop = ((config.height / 2 / scaleFix) >= (startP.top - (deltaY / scaleFix)));
+      const limitBottom = (((config.height / 2 / scaleFix) + (canvas.height) - (startP.top - (deltaY / scaleFix))) <= (config.height / scaleFix));
+
+      // Limit for left
+      if ((limitLeft && !isMinScaleX) || (!limitLeft && isMinScaleX)) {
+        x = startP.x + (startP.left) - (config.width / 2 / scaleFix);
+      }
+
+      // Limit for right
+      if ((limitRight && !isMinScaleX) || (!limitRight && isMinScaleX)) {
+        x = startP.x + (startP.left) + (config.width / 2 / scaleFix) - canvas.width;
+      }
+
+      // Limit for top
+      if ((limitTop && !isMinScaleY) || (!limitTop && isMinScaleY)) {
+        y = startP.y + (startP.top) - (config.height / 2 / scaleFix);
+      }
+
+      // Limit for bottom
+      if ((limitBottom && !isMinScaleY) || (!limitBottom && isMinScaleY)) {
+        y = startP.y + (startP.top) + (config.height / 2 / scaleFix) - canvas.height;
+      }
+
+      // When press shiftKey, deprecated
+      // if (event.srcEvent && event.srcEvent.shiftKey) {
+      //   if (Math.abs(event.deltaX) === Math.max(Math.abs(event.deltaX), Math.abs(event.deltaY))) {
+      //     y = this.offset.top;
+      //   } else {
+      //     x = this.offset.left;
+      //   }
+      // }
+
+      if (x === void 0) { x = (deltaX / scaleFix) + (startP.x); }
+      if (y === void 0) { y = (deltaY / scaleFix) + (startP.y); }
+      this._setStylesForContImg({
+        x, y
+      });
+    }
   }
 
   updatePosition(xOrigin?: number, yOrigin?: number) {
@@ -787,11 +865,7 @@ export class LyImageCropper implements OnDestroy {
       top: originPosition.yc
     };
     this._setStylesForContImg({});
-    this._move({
-      srcEvent: {},
-      deltaX: 0,
-      deltaY: 0
-    });
+    this._simulatePointerMove();
 
     this._cropIfAutoCrop();
   }
@@ -803,57 +877,7 @@ export class LyImageCropper implements OnDestroy {
       config.height / canvas.height);
   }
 
-  private imageSmoothingQuality(img: HTMLCanvasElement, config, quality: number): HTMLCanvasElement {
-    /** Calculate total number of steps needed */
-    let  numSteps = Math.ceil(Math.log(Math.max(img.width, img.height) / Math.max(config.width, config.height)) / Math.log(2)) - 1;
-    numSteps = numSteps <= 0 ? 0 : numSteps;
-
-    /**Array steps */
-    const steps = Array.from(Array(numSteps).keys());
-
-    /** Context */
-    const octx = img.getContext('2d')!;
-
-    const q = ((quality * 10) ** numSteps) / (10 ** numSteps);
-    const fileType = this._defaultType;
-    /** If Steps => imageSmoothingQuality */
-    if (numSteps) {
-      /** Set size */
-      const w = img.width * quality;
-      const h = img.height * quality;
-      /** Only the new img is shown. */
-      if (fileType === 'image/png' || fileType === 'image/svg+xml') {
-        octx.globalCompositeOperation = 'copy';
-      }
-
-      /** Steps */
-      (steps as Array<number>).forEach(() => {
-        octx.drawImage(img,
-          0, 0,
-          w, h
-        );
-      });
-    }
-
-    /**
-     * Step final
-     * Resize & crop image
-     */
-    const oc = document.createElement('canvas'),
-    ctx = oc.getContext('2d')!;
-    oc.width = config.width;
-    oc.height = config.height;
-    ctx.drawImage(img,
-      0, 0,
-      img.width * q, img.height * q,
-      0, 0,
-      oc.width, oc.height
-    );
-    return oc;
-  }
-
   /**
-   * Crop Image
    * Resize & crop image
    */
   crop(config?: ImgCropperConfig): ImgCropperEvent {
@@ -887,13 +911,13 @@ export class LyImageCropper implements OnDestroy {
     ctx.drawImage(this._imgCanvas.nativeElement,
       -(left), -(top),
     );
-    let result = canvasElement;
-    const antiAliasedQ = myConfig.antiAliased ? .5 : 1;
+    const result = canvasElement;
     if (myConfig.output === ImgResolution.Default) {
-      result = this.imageSmoothingQuality(result, config, antiAliasedQ);
+      resizeCanvas(result, config.width, config.height);
     } else if (typeof myConfig.output === 'object') {
-      result = this.imageSmoothingQuality(result, myConfig.output, antiAliasedQ);
+      resizeCanvas(result, myConfig.output.width, myConfig.output.height);
     }
+
     let url: string;
     if (myConfig.type) {
       url = result.toDataURL(`${myConfig.type}`);
@@ -935,6 +959,66 @@ export class LyImageCropper implements OnDestroy {
 
   private _canvaRect(): DOMRect {
     return this._imgCanvas.nativeElement.getBoundingClientRect();
+  }
+
+
+  /** Called when the user has lifted their pointer. */
+  private _pointerUp = (event: TouchEvent | MouseEvent) => {
+    if (this._isSliding) {
+      event.preventDefault();
+      this._removeGlobalEvents();
+      this._isSliding = false;
+      this._startPointerEvent = null;
+    }
+  }
+
+  /** Called when the window has lost focus. */
+  private _windowBlur = () => {
+    // If the window is blurred while dragging we need to stop dragging because the
+    // browser won't dispatch the `mouseup` and `touchend` events anymore.
+    if (this._lastPointerEvent) {
+      this._pointerUp(this._lastPointerEvent);
+    }
+  }
+
+  private _bindGlobalEvents(triggerEvent: TouchEvent | MouseEvent) {
+    const element = this._document;
+    const isTouch = isTouchEvent(triggerEvent);
+    const moveEventName = isTouch ? 'touchmove' : 'mousemove';
+    const endEventName = isTouch ? 'touchend' : 'mouseup';
+    element.addEventListener(moveEventName, this._pointerMove, activeEventOptions);
+    element.addEventListener(endEventName, this._pointerUp, activeEventOptions);
+
+    if (isTouch) {
+      element.addEventListener('touchcancel', this._pointerUp, activeEventOptions);
+    }
+
+    const window = this._getWindow();
+
+    if (typeof window !== 'undefined' && window) {
+      window.addEventListener('blur', this._windowBlur);
+    }
+  }
+
+  /** Removes any global event listeners that we may have added. */
+  private _removeGlobalEvents() {
+    const element = this._document;
+    element.removeEventListener('mousemove', this._pointerMove, activeEventOptions);
+    element.removeEventListener('mouseup', this._pointerUp, activeEventOptions);
+    element.removeEventListener('touchmove', this._pointerMove, activeEventOptions);
+    element.removeEventListener('touchend', this._pointerUp, activeEventOptions);
+    element.removeEventListener('touchcancel', this._pointerUp, activeEventOptions);
+
+    const window = this._getWindow();
+
+    if (typeof window !== 'undefined' && window) {
+      window.removeEventListener('blur', this._windowBlur);
+    }
+  }
+
+  /** Use defaultView of injected document if available or fallback to global window reference */
+  private _getWindow(): Window {
+    return this._document.defaultView || window;
   }
 
 }
@@ -1033,4 +1117,23 @@ function createHtmlImg(src: string) {
   img.crossOrigin = 'anonymous';
   img.src = src;
   return img;
+}
+
+
+function getGesturePointFromEvent(event: TouchEvent | MouseEvent) {
+
+  // `touches` will be empty for start/end events so we have to fall back to `changedTouches`.
+  const point = isTouchEvent(event)
+    ? (event.touches[0] || event.changedTouches[0])
+    : event;
+
+  return {
+    x: point.clientX,
+    y: point.clientY
+  };
+}
+
+/** Returns whether an event is a touch event. */
+function isTouchEvent(event: MouseEvent | TouchEvent): event is TouchEvent {
+  return !!(event as TouchEvent).targetTouches;
 }