+++ src/lib/schematics/ng-add/index.ts
@@ -3,7 +3,7 @@ import { NodePackageInstallTask, RunSchematicTask } from '@angular-devkit/schema
 import { Schema } from './schema';
 import { addPackageToPackageJson, getPackageVersionFromPackageJson } from '../utils/package-config';
 
-const AUI_VERSION = '3.2.8';
+const AUI_VERSION = '3.3.1';
 const HAMMERJS_VERSION = '^2.0.8';
 const ANGULAR_CDK_VERSION = '^9.2.4';